moving to colab

+++ hw3/Part-C/c1.cu
@@ -10,8 +10,8 @@
 #define FW 3
 #define P 1
 
-__global__ void tiledConvolutionKernel(const double* I0, const double* F, double* O, int width, int height) {
-    extern __shared__ double tile[];
+__global__ void tiledConvolutionKernel(const float* I0, const float* F, float* O, int width, int height) {
+    extern __shared__ float tile[];
 
     int x = blockIdx.x * blockDim.x + threadIdx.x;
     int y = blockIdx.y * blockDim.y + threadIdx.y;
@@ -21,7 +21,7 @@ __global__ void tiledConvolutionKernel(const double* I0, const double* F, double
     int threadY = threadIdx.y;
 
     if (x < width && y < height) {
-        double result = 0.0;
+        float result = 0.0f;
         
         for (int c = 0; c < C; ++c) {
             // Load input tile with padding into shared memory
@@ -30,7 +30,7 @@ __global__ void tiledConvolutionKernel(const double* I0, const double* F, double
             if (input_x >= 0 && input_x < width + 2 * P && input_y >= 0 && input_y < height + 2 * P) {
                 tile[threadY * blockDim.x + threadX] = I0[c * (width + 2 * P) * (height + 2 * P) + input_y * (width + 2 * P) + input_x];
             } else {
-                tile[threadY * blockDim.x + threadX] = 0.0;
+                tile[threadY * blockDim.x + threadX] = 0.0f;
             }
 
             __syncthreads();
@@ -56,34 +56,34 @@ __global__ void tiledConvolutionKernel(const double* I0, const double* F, double
     }
 }
 
-void initializeInput(double* I) {
+void initializeInput(float* I) {
     for (int c = 0; c < C; ++c) {
         for (int x = 0; x < H; ++x) {
             for (int y = 0; y < W; ++y) {
-                I[c * H * W + x * W + y] = c * (x + y);
+                I[c * H * W + x * W + y] = static_cast<float>(c * (x + y));
             }
         }
     }
 }
 
-void initializeFilter(double* F) {
+void initializeFilter(float* F) {
     for (int k = 0; k < K; ++k) {
         for (int c = 0; c < C; ++c) {
             for (int i = 0; i < FH; ++i) {
                 for (int j = 0; j < FW; ++j) {
-                    F[k * C * FH * FW + c * FH * FW + i * FW + j] = (c + k) * (i + j);
+                    F[k * C * FH * FW + c * FH * FW + i * FW + j] = static_cast<float>((c + k) * (i + j));
                 }
             }
         }
     }
 }
 
-void addPadding(const double* I, double* I0) {
+void addPadding(const float* I, float* I0) {
     for (int c = 0; c < C; ++c) {
         for (int x = 0; x < H + 2 * P; ++x) {
             for (int y = 0; y < W + 2 * P; ++y) {
                 if (x < P || x >= H + P || y < P || y >= W + P) {
-                    I0[c * (H + 2 * P) * (W + 2 * P) + x * (W + 2 * P) + y] = 0.0;
+                    I0[c * (H + 2 * P) * (W + 2 * P) + x * (W + 2 * P) + y] = 0.0f;
                 } else {
                     I0[c * (H + 2 * P) * (W + 2 * P) + x * (W + 2 * P) + y] = I[c * H * W + (x - P) * W + (y - P)];
                 }
@@ -93,12 +93,12 @@ void addPadding(const double* I, double* I0) {
 }
 
 int main() {
-    size_t imageSize = C * H * W * sizeof(double);
-    size_t paddedSize = C * (H + 2 * P) * (W + 2 * P) * sizeof(double);
-    size_t filterSize = K * C * FH * FW * sizeof(double);
-    size_t outputSize = K * H * W * sizeof(double);
+    size_t imageSize = C * H * W * sizeof(float);
+    size_t paddedSize = C * (H + 2 * P) * (W + 2 * P) * sizeof(float);
+    size_t filterSize = K * C * FH * FW * sizeof(float);
+    size_t outputSize = K * H * W * sizeof(float);
 
-    double *I, *I0, *F, *O;
+    float *I, *I0, *F, *O;
     cudaMallocManaged(&I, imageSize);
     cudaMallocManaged(&I0, paddedSize);
     cudaMallocManaged(&F, filterSize);
@@ -110,7 +110,7 @@ int main() {
 
     dim3 blockSize(16, 16);
     dim3 gridSize((W + blockSize.x - 1) / blockSize.x, (H + blockSize.y - 1) / blockSize.y, K);
-    size_t sharedMemSize = blockSize.x * blockSize.y * sizeof(double);
+    size_t sharedMemSize = blockSize.x * blockSize.y * sizeof(float);
 
     auto start = std::chrono::high_resolution_clock::now();
     tiledConvolutionKernel<<<gridSize, blockSize, sharedMemSize>>>(I0, F, O, W, H);
@@ -120,7 +120,7 @@ int main() {
     std::chrono::duration<double> elapsed = end - start;
     std::cout << "Kernel execution time: " << elapsed.count() << " seconds" << std::endl;
 
-    double checksum = 0.0;
+    float checksum = 0.0f;
     for (int i = 0; i < K * W * H; ++i) {
         checksum += O[i];
     }
